Implemented saving the results to a database. Can now run the module from the run_all.py script.
Fixed a small typo in the documentation for store_flooded_buildings_in_database which defined flood_model_id as a float. Now changed to an int.
